Pulls out `useFilters` hook
This is almost ready for reuse!

Despite my best intentions this wasn't quite a 1:1 change as I found a
few small bugs after I ported it on over
